moved Statuses outside of BatchStatusService
